Extend Content Module
# add more funcitons to the content helper module
# change the corresponding code in the other files
# translate the search header (i18n)
